+++ applications/system/sl_updater/fkermit.c
