+++ applications/system/sl_updater/sl_update_cli.c
@@ -1,4 +1,4 @@
-#include "sl_update.h"
+#include "sl_updater.h"
 
 #include <furi.h>
 #include <cli/cli.h>
