+++ applications/system/sl_updater/sl_updater.c
@@ -1,4 +1,4 @@
-#include "sl_update.h"
+#include "sl_updater.h"
 
 #include <furi.h>
